requirements.txt
openai==0.27.9
requests==2.31.0
pillow==10.0.0
